nyayasahayak_backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.119.1"}
psycopg = {extras = ["binary"], version = "^3.2.11"}
pandas = "^2.3.3"
pyarrow = "^21.0.0"
chromadb = "^1.2.1"
langchain = "^1.0.2"
langchain-community = "^0.4"
langchain-openai = "^1.0.1"
langgraph = "^1.0.1"
python-dotenv = "^1.1.1"
sentence-transformers = "^5.1.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
